fix: change layout integration and some design improvements
