+++ components/layouts/main-layout.tsx
@@ -1,0 +1,18 @@
+import Head from "next/head";
+import { Header } from "../header";
+import { LayoutProps } from "./types/page-with-layout";
+
+const MainLayout: LayoutProps = ({ children }) => {
+  return (
+    <>
+      <Head>
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <Header />
+      <div className="w-screen h-screen bg-slate-100">
+        <div className="w-full sm:w-7/12 px-s my-0 mx-auto">{children}</div>
+      </div>
+    </>
+  );
+};
+export default MainLayout;
