+++ components/layouts/types/page-with-layout.tsx
@@ -1,0 +1,7 @@
+import type { ReactElement } from "react";
+
+export type LayoutProps = ({
+  children,
+}: {
+  children: ReactElement;
+}) => ReactElement;
