+++ pages/_app.tsx
@@ -1,9 +1,7 @@
 import { SessionProvider } from "next-auth/react";
 import type { AppProps } from "next/app";
 import { Poppins } from "@next/font/google";
-import { PageWithLayoutType } from "../components/layouts/types/PageWithLayout";
 import "../styles/globals.css";
-import { ReactElement } from "react";
 
 const poppins = Poppins({
   subsets: ["latin"],
@@ -12,22 +10,14 @@ const poppins = Poppins({
   variable: "--font-poppins",
 });
 
-type AppLayoutProps = AppProps & {
-  Component: PageWithLayoutType;
-};
-
 export default function App({
   Component,
   pageProps: { session, ...pageProps },
-}: AppLayoutProps) {
-  const Layout =
-    Component.layout || ((children: ReactElement) => <>{children}</>);
+}: AppProps) {
   return (
     <SessionProvider session={session}>
       <main className={`${poppins.variable} font-poppins`}>
-        <Layout>
-          <Component {...pageProps} />
-        </Layout>
+        <Component {...pageProps} />
       </main>
     </SessionProvider>
   );
