+++ pages/index.tsx
@@ -1,6 +1,8 @@
 import { GetServerSideProps, InferGetStaticPropsType } from "next";
-import MainLayout from "../components/layouts/MainLayout";
+import MainLayout from "../components/layouts/main-layout";
 import {
+  Heading,
+  HeadingSize,
   Stack,
   StackDirection,
   StackSpacing,
@@ -15,23 +17,34 @@ export default function PageHome({}: PageProps): InferGetStaticPropsType<
   typeof getServerSideProps
 > {
   return (
-    <>
-      <Head>
-        <title>Mumble Home</title>
-      </Head>
-      <Stack
-        direction={StackDirection.col}
-        spacing={StackSpacing.s}
-        withDivider={true}
-      >
-        <div>Mumble 1</div>
-        <div>Mumble 2</div>
-      </Stack>
-    </>
+    <MainLayout>
+      <>
+        <Head>
+          <title>Mumble Home</title>
+        </Head>
+        <div className="text-violet-600 pt-l">
+          <Heading headingLevel={HeadingSize.h1}>Willkommen auf Mumble</Heading>
+        </div>
+        <div className="text-slate-500 pt-xs pb-l">
+          <Heading headingLevel={HeadingSize.h4}>
+            Voluptatem qui cumque voluptatem quia tempora dolores distinctio vel
+            repellat dicta.
+          </Heading>
+        </div>
+        <div className="bg-white">
+          <Stack
+            direction={StackDirection.col}
+            spacing={StackSpacing.s}
+            withDivider={true}
+          >
+            <div>Mumble 1</div>
+            <div>Mumble 2</div>
+          </Stack>
+        </div>
+      </>
+    </MainLayout>
   );
 }
-
-PageHome.layout = MainLayout;
 
 export const getServerSideProps: GetServerSideProps = async () => ({
   props: { posts: require("../data/posts.json") },
